pyproject.toml
[project]
name = "display-auto-rotation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
]

[project.scripts]
display-auto-rotation = "display_auto_rotation.main:main"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
